requirements.txt
requests==2.28.0
python-telegram-bot==13.12
Pillow==9.1.1
torch==1.11.0
torchvision==0.12.0
resnet
tensorflow
numpy